write: Correctly handle huge write requests
(cherry picked from commit e26d91cef1272f445d43fc70b2b4ec7c94a63185)

+++ libarchive/archive_write_private.h
@@ -69,7 +69,7 @@ void __archive_write_filters_free(struct archive *);
 struct archive_write_filter *__archive_write_allocate_filter(struct archive *);
 
 int __archive_write_output(struct archive_write *, const void *, size_t);
-int __archive_write_nulls(struct archive_write *, size_t);
+int __archive_write_nulls(struct archive_write *, uint64_t);
 int __archive_write_filter(struct archive_write_filter *, const void *, size_t);
 
 struct archive_write {
